requirements.txt
google-analytics-data==0.18.19
google-analytics-admin==0.24.1
google-auth~=2.40
mcp[cli]>=1.2.0
httpx>=0.28.1